% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.records.R
\name{aggregate_records}
\alias{aggregate_records}
\title{Aggregate database records}
\usage{
aggregate_records(x, ...)
}
\arguments{
\item{x}{A record set}

\item{...}{A set of database field names used in the aggregation process}
}
\value{
A dataframe summarizing counts of prey items in snake diet samples.
See details for how summarization takes place.
}
\description{
Aggregate database records according to different criteria
}
\details{
The default result is a three column dataframe where each row is
a tuple of the form (n, p, q), where n is a snake species, p is a prey group,
and q = min(predator_count, prey_count). In other words, each row is the
minimum number of instances of prey group p recorded in the diet of snake
taxon n. Additional database fields specified via \code{...} augment this
row tuple with more information. For example, passing "locality_adm0_name"
through \code{...} would yield tuples of the form (n, p, r, q), and the
interpretation becomes the minimum number of instances of prey group p
recorded in the diet of snake taxon n in country r.
}
