% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.records.R
\name{collapse_ranks}
\alias{collapse_ranks}
\title{Collapse predator ranks}
\usage{
collapse_ranks(x, predator_rank)
}
\arguments{
\item{x}{A record set}

\item{predator_rank}{Desired taxonomic level. Should be one of "species",
"genus", or "family"}
}
\value{
A record set with predator ranks collapsed to the desired level.
}
\description{
Collapse predator ranks to desired taxonomic level
}
