% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.records.R
\name{filter_records}
\alias{filter_records}
\title{Filter database records}
\usage{
filter_records(x, ...)
}
\arguments{
\item{x}{A record set (optional). If ommitted defaults to the full database.}

\item{...}{Filter criteria. Valid values can be any database field (for a
description of database fields see \code{?diet}) or the values xmin,
xmax, ymin, ymax.}
}
\value{
A record set (a set of rows in the database) where each row matches
the criteria passed to the function via the \code{...} parameter.
}
\description{
Filter food records by taxonomic and geographic criteria
}
