% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.records.R
\name{group_prey}
\alias{group_prey}
\title{Classify prey into groups}
\usage{
group_prey(x, group = "coarse")
}
\arguments{
\item{x}{A record set}

\item{group}{Either a character vector specifying one of the built-in prey
classifications ("coarse" or "detailed") or a user-defined function modeled
after these two built-ins that will perform a custom prey classification.}
}
\value{
A record set with an additional column -- "prey_group" -- containing
the result of the prey categorization.
}
\description{
Classify prey into a set of user-defined categories
}
\seealso{
\code{\link{prey_coarse}}, \code{\link{prey_detailed}}
}
