% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food.records.R
\name{prey_coarse}
\alias{prey_coarse}
\alias{prey_detailed}
\title{Built-in prey classification schemes}
\usage{
prey_coarse(x)

prey_detailed(x)
}
\arguments{
\item{x}{A record set.}
}
\value{
A character vector naming the prey group of each row in the record
set.
}
\description{
Categorize prey items according to higher taxonomy
}
\details{
These two functions inspect the higher taxonomy and lifestage
fields of each row in the record set and automatically make a determination
as to which prey group the prey item belongs to. The functions differ in
their level of taxonomic resolution. The different prey groups and criteria
used to perform the categorizations are accessible by inspecting the function
body, which can be achieved by typing the function name into the R console.
}
\section{Functions}{
\itemize{
\item \code{prey_detailed}: prey_detailed
}}

