% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squamatabase.R
\docType{package}
\name{squamatabase}
\alias{squamatabase}
\alias{squamatabase-package}
\title{squamatabase: a database of diet samples from the world's snakes}
\description{
For a description of the database see `?diet` or `help(diet)`
}
\section{Functions}{

filter_records
group_prey
aggregate_records
prey_coarse
prey_detailed
}

